Move things around to make the code easier to read
